requirements.txt
langchain 
langchain-core
langchain-community
langchain-groq
langchain-text-splitters
streamlit
python-dotenv
setuptools
pillow
langchain-tavily

